chore: change name to Auth service
