--- account-service/src/main/java/com/account/controller/AccountController.java
@@ -1,26 +1,0 @@
-package com.account.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.account.vo.ApiResponse;
-
-@RestController
-@RequestMapping("/auth")
-public class AccountController {
-
-    @PostMapping("/register")
-	public ResponseEntity<ApiResponse<String>> signIn() {
-		ApiResponse<String> response = new ApiResponse<>(HttpStatus.OK.value(), "successed", "signUp");
-		return new ResponseEntity<>(response, HttpStatus.OK);
-	}
-
-	@PostMapping("/login")
-	public ResponseEntity<ApiResponse<String>> signUp() {
-		ApiResponse<String> response = new ApiResponse<>(HttpStatus.OK.value(), "successed", "signIn");
-		return new ResponseEntity<>(response, HttpStatus.OK);
-	}
-}
