--- account-service/src/main/java/com/account/mapper/UserMapper.java
@@ -1,30 +1,0 @@
-package com.account.mapper;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Delete;
-import org.apache.ibatis.annotations.Insert;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Select;
-import org.apache.ibatis.annotations.Update;
-
-import com.account.entity.User;
-
-@Mapper
-public interface UserMapper {
-
-    @Select("SELECT * FROM users WHERE id = #{id}")
-    User getUserById(Long id);
-
-    @Select("SELECT * FROM users")
-    List<User> getAllUsers();
-
-    @Insert("INSERT INTO users (name, email) VALUES (#{name}, #{email})")
-    void insertUser(User user);
-
-    @Update("UPDATE users SET name=#{name}, email=#{email} WHERE id=#{id}")
-    void updateUser(User user);
-
-    @Delete("DELETE FROM users WHERE id = #{id}")
-    void deleteUser(Long id);
-}
