--- account-service/src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,16 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: 'account-service-1'
-      author: Emberow
-      changes:
-        - sqlFile:
-            encoding: utf8
-            path: classpath:db/changelog/sqlfile/20240626_1.sql
-
-  - changeSet:
-      id: 'account-service-2'
-      author: Emberow
-      changes:
-        - sqlFile:
-            encoding: utf8
-            path: classpath:db/changelog/sqlfile/20240626_2.sql
