+++ auth-service/.editorconfig
