+++ auth-service/src/main/java/com/auth/AuthService.java
@@ -1,0 +1,18 @@
+package com.auth;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.metrics.buffering.BufferingApplicationStartup;
+
+@SpringBootApplication
+public class AuthService {
+
+        private static final int CAPACITY = 2048;
+        public static void main(String[] args) {
+                SpringApplication application = new SpringApplication(AuthService.class);
+                application.setApplicationStartup(new BufferingApplicationStartup(CAPACITY));
+                application.run(args);
+                System.out.println("٩(๑❛ᴗ❛๑)۶  Auth Service 啟動!!  ٩(๑❛ᴗ❛๑)۶");
+        }
+
+}
