+++ auth-service/src/main/java/com/auth/service/HelloService.java
@@ -1,0 +1,11 @@
+package com.auth.service;
+
+import com.auth.vo.HelloRequestVO;
+
+public interface HelloService {
+
+    String echo(String input);
+
+    String hello(HelloRequestVO helloRequestVO);
+}
+
