+++ auth-service/src/test/java/com/auth/AuthServiceTests.java
@@ -1,0 +1,13 @@
+package com.auth;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = AuthService.class)
+class AuthServiceTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
